quantum_cdf_wavelet.py
import numpy as np
import matplotlib.pyplot as plt
from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister
from qiskit import Aer, execute, transpile
from qiskit.visualization import plot_histogram, circuit_drawer
from qiskit.circuit.library import QFT
from qiskit.quantum_info import Statevector
import cv2
from PIL import Image
import math

class QuantumCDFWaveletTransform:
    """
    基于CDF(2,2)小波变换的量子提升方案实现
    
    实现论文中的三个步骤：
    1. Split: 分离奇偶样本
    2. Predict: P(S) = 1/2[S(2i) + S(2i+2)]
    3. Update: W(D) = 1/4[D(i-1) + D(i)]
    """
    
    def __init__(self, bit_precision=4):
        self.bit_precision = bit_precision
        self.max_value = 2**bit_precision - 1
        
    def create_split_circuit(self, signal_length):
        """
        创建Split步骤的量子电路
        将信号S分离为奇偶样本
        """
        # 需要足够的量子比特来表示信号
        n_qubits = int(np.ceil(np.log2(signal_length))) + self.bit_precision
        
        # 创建量子寄存器
        signal_reg = QuantumRegister(self.bit_precision, 'signal')
        index_reg = QuantumRegister(int(np.ceil(np.log2(signal_length))), 'index')
        even_reg = QuantumRegister(self.bit_precision, 'even')
        odd_reg = QuantumRegister(self.bit_precision, 'odd')
        
        # 经典寄存器用于测量
        cr = ClassicalRegister(self.bit_precision * 2, 'result')
        
        qc = QuantumCircuit(signal_reg, index_reg, even_reg, odd_reg, cr)
        
        # 添加分离逻辑 - 基于索引的奇偶性
        # 使用控制门来实现条件分离
        for i in range(len(index_reg)):
            # 如果索引的最低位为0（偶数），复制到even_reg
            qc.x(index_reg[0])  # 翻转最低位
            for j in range(self.bit_precision):
                qc.ccx(index_reg[0], signal_reg[j], even_reg[j])
            qc.x(index_reg[0])  # 恢复
            
            # 如果索引的最低位为1（奇数），复制到odd_reg
            for j in range(self.bit_precision):
                qc.ccx(index_reg[0], signal_reg[j], odd_reg[j])
        
        return qc
    
    def create_predict_circuit(self):
        """
        创建Predict步骤的量子电路
        实现 P(S) = 1/2[S(2i) + S(2i+2)]
        """
        # 量子寄存器
        s_2i = QuantumRegister(self.bit_precision, 's_2i')        # S(2i)
        s_2i_plus_2 = QuantumRegister(self.bit_precision, 's_2i+2')  # S(2i+2)
        predict_reg = QuantumRegister(self.bit_precision + 1, 'predict')  # P(S)，需要额外一位处理进位
        
        cr = ClassicalRegister(self.bit_precision + 1, 'predict_result')
        
        qc = QuantumCircuit(s_2i, s_2i_plus_2, predict_reg, cr)
        
        # 实现量子加法器：predict_reg = (s_2i + s_2i_plus_2) / 2
        # 首先实现加法
        self._quantum_adder(qc, s_2i, s_2i_plus_2, predict_reg)
        
        # 然后除以2（右移一位）
        for i in range(self.bit_precision):
            if i < self.bit_precision - 1:
                qc.cx(predict_reg[i+1], predict_reg[i])
        
        return qc
    
    def create_update_circuit(self):
        """
        创建Update步骤的量子电路
        实现 W(D) = 1/4[D(i-1) + D(i)]
        """
        # 量子寄存器
        d_i_minus_1 = QuantumRegister(self.bit_precision, 'd_i-1')  # D(i-1)
        d_i = QuantumRegister(self.bit_precision, 'd_i')            # D(i)
        update_reg = QuantumRegister(self.bit_precision + 2, 'update')  # W(D)，需要额外两位
        
        cr = ClassicalRegister(self.bit_precision + 2, 'update_result')
        
        qc = QuantumCircuit(d_i_minus_1, d_i, update_reg, cr)
        
        # 实现量子加法器：update_reg = (d_i_minus_1 + d_i) / 4
        # 首先实现加法
        self._quantum_adder(qc, d_i_minus_1, d_i, update_reg)
        
        # 然后除以4（右移两位）
        for i in range(self.bit_precision):
            if i < self.bit_precision - 2:
                qc.cx(update_reg[i+2], update_reg[i])
        
        return qc
    
    def _quantum_adder(self, qc, reg_a, reg_b, result_reg):
        """
        实现量子加法器
        result_reg = reg_a + reg_b
        """
        # 使用量子傅里叶变换实现加法
        # 简化版本：使用控制门实现基本加法
        carry = 0
        
        for i in range(len(reg_a)):
            # 半加器逻辑
            qc.cx(reg_a[i], result_reg[i])
            qc.cx(reg_b[i], result_reg[i])
            
            # 进位逻辑
            if i < len(reg_a) - 1:
                qc.ccx(reg_a[i], reg_b[i], result_reg[i+1])
    
    def create_complete_cdf_circuit(self, signal_data):
        """
        创建完整的CDF(2,2)小波变换量子电路
        """
        signal_length = len(signal_data)
        
        # 创建主要的量子寄存器
        signal_reg = QuantumRegister(self.bit_precision, 'signal')
        index_reg = QuantumRegister(int(np.ceil(np.log2(signal_length))), 'index')
        
        # Split步骤的寄存器
        even_reg = QuantumRegister(self.bit_precision, 'even')
        odd_reg = QuantumRegister(self.bit_precision, 'odd')
        
        # Predict步骤的寄存器
        predict_reg = QuantumRegister(self.bit_precision + 1, 'predict')
        detail_reg = QuantumRegister(self.bit_precision, 'detail')
        
        # Update步骤的寄存器
        approx_reg = QuantumRegister(self.bit_precision, 'approx')
        
        # 经典寄存器
        cr = ClassicalRegister(self.bit_precision * 3, 'result')
        
        qc = QuantumCircuit(signal_reg, index_reg, even_reg, odd_reg, 
                           predict_reg, detail_reg, approx_reg, cr)
        
        # 初始化信号数据
        for i, value in enumerate(signal_data):
            # 将每个信号值编码到量子态中
            binary_value = format(int(value), f'0{self.bit_precision}b')
            for j, bit in enumerate(binary_value):
                if bit == '1':
                    qc.x(signal_reg[j])
        
        # Step 1: Split - 分离奇偶样本
        qc.barrier(label='Split')
        # 实现分离逻辑...
        
        # Step 2: Predict - 计算预测值
        qc.barrier(label='Predict')
        # 实现预测逻辑...
        
        # Step 3: Update - 计算更新值
        qc.barrier(label='Update')
        # 实现更新逻辑...
        
        return qc
    
    def classical_cdf_transform(self, signal):
        """
        经典CDF(2,2)小波变换实现，用于对比验证
        """
        signal = np.array(signal, dtype=float)
        n = len(signal)
        
        # Step 1: Split
        even = signal[::2]  # S(2i)
        odd = signal[1::2]  # S(2i+1)
        
        # Step 2: Predict - P(S) = 1/2[S(2i) + S(2i+2)]
        predict = np.zeros_like(odd)
        for i in range(len(odd)):
            if i == 0:
                # 边界处理
                predict[i] = even[0]
            elif i == len(odd) - 1:
                # 边界处理
                predict[i] = even[-1]
            else:
                predict[i] = 0.5 * (even[i] + even[i+1])
        
        # 计算详细系数 D(i) = S(2i+1) - P(S)
        detail = odd - predict
        
        # Step 3: Update - W(D) = 1/4[D(i-1) + D(i)]
        update = np.zeros_like(even)
        for i in range(len(even)):
            if i == 0:
                # 边界处理
                if len(detail) > 0:
                    update[i] = 0.25 * detail[0]
            elif i == len(even) - 1:
                # 边界处理
                if len(detail) > i-1:
                    update[i] = 0.25 * detail[i-1]
            else:
                if i-1 < len(detail) and i < len(detail):
                    update[i] = 0.25 * (detail[i-1] + detail[i])
        
        # 计算近似系数 A(i) = S(2i) + W(D)
        approx = even + update
        
        return approx, detail
    
    def image_to_signal(self, image_path, block_size=(2, 2)):
        """
        将图像转换为信号块进行处理
        """
        # 读取图像
        if isinstance(image_path, str):
            image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
        else:
            image = image_path
            
        if image is None:
            # 创建测试图像
            image = np.array([
                [1, 3, 2, 1],
                [2, 6, 6, 6],
                [5, 2, 7, 4],
                [3, 1, 7, 2]
            ], dtype=np.uint8)
        
        # 将图像分割为块
        h, w = image.shape
        blocks = []
        
        for i in range(0, h, block_size[0]):
            for j in range(0, w, block_size[1]):
                block = image[i:i+block_size[0], j:j+block_size[1]]
                blocks.append(block.flatten())
        
        return blocks, image
    
    def visualize_transform_results(self, original_signal, approx, detail):
        """
        可视化变换结果
        """
        fig, axes = plt.subplots(3, 1, figsize=(12, 8))
        
        # 原始信号
        axes[0].plot(original_signal, 'b-o', label='Original Signal')
        axes[0].set_title('Original Signal')
        axes[0].set_xlabel('Sample Index')
        axes[0].set_ylabel('Amplitude')
        axes[0].grid(True)
        axes[0].legend()
        
        # 近似系数
        axes[1].plot(approx, 'r-s', label='Approximation Coefficients')
        axes[1].set_title('Approximation Coefficients (Low-pass)')
        axes[1].set_xlabel('Sample Index')
        axes[1].set_ylabel('Amplitude')
        axes[1].grid(True)
        axes[1].legend()
        
        # 详细系数
        axes[2].plot(detail, 'g-^', label='Detail Coefficients')
        axes[2].set_title('Detail Coefficients (High-pass)')
        axes[2].set_xlabel('Sample Index')
        axes[2].set_ylabel('Amplitude')
        axes[2].grid(True)
        axes[2].legend()
        
        plt.tight_layout()
        return fig
    
    def run_quantum_simulation(self, signal_data):
        """
        运行量子电路模拟
        """
        # 创建完整电路
        qc = self.create_complete_cdf_circuit(signal_data)
        
        # 使用Aer模拟器
        simulator = Aer.get_backend('qasm_simulator')
        
        # 添加测量
        qc.measure_all()
        
        # 编译和执行
        compiled_circuit = transpile(qc, simulator)
        job = execute(compiled_circuit, simulator, shots=1024)
        result = job.result()
        
        return result, qc