requirements.txt
qiskit>=0.45.0
qiskit-aer>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.7.0